+++ libft/ft_memcmp.c
@@ -6,11 +6,11 @@
 /*   By: zoyern <zoyern@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 17:29:18 by zoyern            #+#    #+#             */
-/*   Updated: 2023/11/03 18:59:35 by zoyern           ###   ########.fr       */
+/*   Updated: 2023/11/03 23:14:08 by zoyern           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+int	ft_memcmp(const void *s1, const void *s2, unsigned long n)
 {
 	const unsigned char	*tmp_s1;
 	const unsigned char	*tmp_s2;
