+++ libft/ft_memmove.c
@@ -6,11 +6,11 @@
 /*   By: zoyern <zoyern@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 00:38:03 by almounib          #+#    #+#             */
-/*   Updated: 2023/11/02 10:22:46 by zoyern           ###   ########.fr       */
+/*   Updated: 2023/11/03 23:14:08 by zoyern           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, const void *src, unsigned int n)
+void	*ft_memmove(void *dest, const void *src, unsigned long n)
 {
 	unsigned char	*tmp_dest;
 	unsigned char	*tmp_src;
